refactor code to get ArtPieces component, adjust img size, add h1 to homepage
